requirements.txt
pandas
numpy 
yfinance 
pyaml 
hopsworks[python] 
python-dotenv
mlflow
dagshub
xgboost

